requirements.txt
# Flask Web Framework and Extensions
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1

# Database
pymongo==4.5.0

# Environment Configuration
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Data Processing
pandas==2.1.1
numpy==1.24.3

# Image Processing and OCR
img2table==1.2.4
paddlepaddle==2.5.1
paddleocr==2.7.0.3
opencv-python==4.8.1.78
Pillow==10.0.1

# Additional Dependencies
bson==0.5.10

# SMS/OTP Service
twilio==8.10.0